requirements.txt
colorama==0.4.6
pydantic==2.11.3
mcp-cli==0.1.0